Исправляет структуру проекта по результатам замечаний наставника
