--- src/components/filter.js
@@ -1,12 +1,0 @@
-const filterNames = [`all`, `overdue`, `today`, `favorites`, `repeating`, `archive`];
-
-const generateFilter = () => {
-  return filterNames.map((it) => {
-    return {
-      title: it,
-      count: Math.floor(Math.random() * 10),
-    };
-  });
-};
-
-export {generateFilter};
