+++ src/components/filter/create-filters-template.js
@@ -15,6 +15,7 @@ const createFilterMarkup = (filter, isChecked) => {
 
 const createFiltersTemplate = (filters) => {
   const filtersMarkup = filters.map((it, i) => createFilterMarkup(it, i === 0)).join(`\n`);
+
   return `<section class="main__filter filter container">
             ${filtersMarkup}
           </section>`;
