--- src/components/task.js
@@ -1,64 +1,0 @@
-import {COLORS} from "./const.js";
-
-const DescriptionItems = [
-  `Изучить теорию`,
-  `Сделать домашку`,
-  `Пройти интенсив на соточку`,
-];
-
-const DefaultRepeatingDays = {
-  "mo": false,
-  "tu": false,
-  "we": false,
-  "th": false,
-  "fr": false,
-  "sa": false,
-  "su": false,
-};
-
-const getRandomArrayItem = (array) => {
-  const randomIndex = getRandomIntegerNumber(0, array.length);
-
-  return array[randomIndex];
-};
-
-const getRandomIntegerNumber = (min, max) => {
-  return min + Math.floor(Math.random() * (max - min));
-};
-
-const getRandomDate = () => {
-  const targetDate = new Date();
-  const sign = Math.random() > 0.5 ? 1 : -1;
-  const diffValue = sign * getRandomIntegerNumber(0, 8);
-
-  targetDate.setDate(targetDate.getDate() + diffValue);
-
-  return targetDate;
-};
-
-const generateRepeatingDays = () => {
-  const dayKeys = Object.keys(DefaultRepeatingDays);
-  const repeatingDay = dayKeys.map((day) => ({[day]: Math.random() > 0.5}));
-  return Object.assign({}, ...repeatingDay);
-};
-
-const generateTask = () => {
-  const dueDate = Math.random() > 0.5 ? null : getRandomDate();
-
-  return {
-    description: getRandomArrayItem(DescriptionItems),
-    dueDate,
-    repeatingDays: dueDate ? DefaultRepeatingDays : generateRepeatingDays(),
-    color: getRandomArrayItem(COLORS),
-    isArchive: Math.random() > 0.5,
-    isFavorite: Math.random() > 0.5,
-  };
-};
-
-const generateTasks = (count) => {
-  return new Array(count)
-    .fill(``)
-    .map(generateTask);
-};
-
-export {generateTask, generateTasks};
