+++ src/components/task/create-colors-markup.js
@@ -1,0 +1,23 @@
+const createColorsMarkup = (colors, currentColor) => {
+  return colors
+    .map((color, index) => {
+      return (
+        `<input
+          type="radio"
+          id="color-${color}-${index}"
+          class="card__color-input card__color-input--${color} visually-hidden"
+          name="color"
+          value="${color}"
+          ${currentColor === color ? `checked` : ``}
+        />
+        <label
+          for="color-${color}--${index}"
+          class="card__color card__color--${color}"
+          >${color}
+        </label>`
+      );
+    })
+    .join(`\n`);
+};
+
+export {createColorsMarkup};
