+++ src/main.js
@@ -1,12 +1,12 @@
-import {createFiltersTemplate} from "./components/filters-template.js";
-import {createMenuTemplate} from "./components/menu-template.js";
-import {createSortTemplate} from "./components/sort-template.js";
-import {createTaskEditTemplate} from "./components/task-edit-template.js";
-import {createTaskTemplate} from "./components/task-template.js";
-import {renderBoardContainer} from "./components/board.js";
-import {renderButtonLoadMore} from "./components/button-load-more.js";
-import {generateFilter} from "./components/filter.js";
-import {generateTasks} from "./components/task.js";
+import {createFiltersTemplate} from "./components/filter/create-filters-template.js";
+import {createMenuTemplate} from "./components/menu/create-menu-template.js";
+import {createSortTemplate} from "./components/sort/create-sort-template.js";
+import {createTaskEditTemplate} from "./components/task/create-task-edit-template.js";
+import {createTaskTemplate} from "./components/task/create-task-template.js";
+import {renderBoardContainer} from "./components/board/render-board-container.js";
+import {renderButtonLoadMore} from "./components/button-load-more/render-button-load-more.js";
+import {generateFilter} from "./components/filter/generate-filters.js";
+import {generateTasks} from "./components/task/generate-task.js";
 
 const TASK_COUNT = 22;
 const SHOWING_TASKS_COUNT_ON_START = 8;
