+++ src/mocks/description-items.js
@@ -1,0 +1,7 @@
+const DescriptionItems = [
+  `Изучить теорию`,
+  `Сделать домашку`,
+  `Пройти интенсив на соточку`,
+];
+
+export {DescriptionItems};
