+++ src/mocks/get-random-date.js
@@ -1,0 +1,13 @@
+import {getRandomIntegerNumber} from '../helpers/utils.js';
+
+const getRandomDate = () => {
+  const targetDate = new Date();
+  const sign = Math.random() > 0.5 ? 1 : -1;
+  const diffValue = sign * getRandomIntegerNumber(0, 8);
+
+  targetDate.setDate(targetDate.getDate() + diffValue);
+
+  return targetDate;
+};
+
+export {getRandomDate};
